Optimize `is_square_safe` (?)
Instead of checking every square if it threatens a target square, we
instead check if a target square could be threatened by the squares that
could potentially threaten it. In other words, from the target tile, we
check it's rank and file for rooks/queens, its diagonals for bishops/queens,
knights moves for knights, and so on.