Implement alpha-beta pruning and also use the killer move heuristics
This provides our program with a near 12x speedup as our pruning
lets us consider less branches over all and makes us search less branches
as well! With this, we can search around 50k to 100k branches per second.

Alpha Beta pruning works by realizing that it is impossible for us to
do better than the worst possible outcome that the enemy player can force
us to get, and that we shouldn't even both to consider branches that
are worse than the best possible outcome we can achieve. In other words,
alpha is a lower bound and beta is an upper bound.

Suppose that we have some branch. We know we will always pick the "best"
scoring leaf we can, and that the enemy will always pick the "worst" scoring
leaf it can (worse score for us = better score for them). Therefore, if the
enemy can always make us get a score of at most `beta`, then there is no
reason to try and check out a branch that would end up with a score higher,
because the enemy would just instead pick the move that would make our score
at most beta. Similarly, if we know we can always achieve a score of at least
alpha, then there is no reason to try and check out a branch that would end
up with a score of less than alpha, because we could instead just play the
better branch.

Another way to look at this is to note how `max(9, min(5, ...))` has to equal
at least 9, because `min(5, ...)` has to equal at most 5, and 5 < 9, so we
don't actually need to look at the rest of `min`'s values.

The killer move heurstic is an improvement for alpha beta pruning. Since
alpha and beta are initially negative and positive infinity and gradually
narrow as the search goes on, it would be nice if we could start with an
alpha and beta that are close to the alpha and beta we had at the end last
time, so that we don't need to recompute it and also so that we can
get the benefit of that narrower window. We do this by saving the best two
moves found at each depth of the tree, and hope that next turn still
allows us to play the move. We keep two here because sometimes the
enemy player will make a move that prevents us from doning of the moves,
so we keep track of two to minimize that posibility.