Improve position evaluation with position tables.
Position tables are an easy way to get the AI to try to develop its pieces
We give a bonus or penalty for placing certain pieces on certain squares.
For example, we might give a penalty for placing the king out in the open,
or a bonus for placing the knight in the center, since that's usually a
better place to put them.

More advanced evaluation metrics might include checking if a piece is threatening
a particular square/enemy piece, or evaluating the quality of the pawn structure.