Implement iterative deepening
Iterative deepening makes our AI search at smaller depths initially
before searching at the full depth. This is useful because it allows for
the priniple variation (and other heuristics once implemented) to "build
up" and provide a good move order for when we finally do search at the full
depth. This causes us to redo our work, but saves time in the long run
because we are able to prune much more agressively with the improved
move ordering.