Create Selector to allow for human/AI difficulty level selection.
The Selector stores a bunch of buttons and keeps track of which one was
last pressed ("selected"). This is used to let the player select
the difficulty of the AI they want to face against.

At the moment, Selector is a Layout, but I think it doesn't need to be
(ideally, I want any number of Buttons to be Selector-able, but maybe I
want to lay them out in different ways. Ultimately, the Selector shouldn't
care about that stuff, just that it knows that it has a bunch of buttons.)