Create "BasicAction" enum
A "basic action" is meant to break down a complex player action (such
as a castle) into a set of simpler actions (a castle is really just two
moves at the same time). This should make it easier to work with at the
UI level since we really only care about moving or removing pieces and not
if the move happens to be a lunge or normal move or capture.